[#79] test : 선착순 쿠폰 생성 로직 변경
